scripts/fips_crypto_hmac.sh: Unbreak FIPS with Bash 5
FIPS (CONFIG_CRYPTO_FIPS) compares a build-time and runtime digest
during initialization as part of its integrity check. However, the
script generating the build-time digest relied on iteration order of
bash associative arrays, which changed in newer Bash versions, causing
the digest to change. This caused features which relied on FIPS, such
as Full Device Encryption, to fail and cause a kernel panic.

Use a regular array instead of an associative array to fix the
iteration order.